fix: read compressed integers properly.
Compressed integers were being read as unsigned VLQ which is not
correct. The encoding isn't exactly LEB128 either, it's very close to
unsigned LEB128, but the end result may be signed.

The way it works is as follows:
- For 64 bit integers, unsigned LEB128 encoding is used, with a
difference: the encoded number is limited to 9 bytes, and the in 9th
byte all the 8 bits are encoded as is (so the 64 bit include 8
bytes  * 7 bits + 1 byte * 8 bits). For negative integers, the MSB
will be set to 1.
- For 16 and 32 bit integers, the same encoding is used, except the
only the less significant 16 and 32 bits respectively are
non-zero. The resulting numbers are also signed and can be negative.